+++ lib/controller.js
@@ -12,18 +12,20 @@ var usb = require('./usb_connection')
 */
 function listTessels(callback) {
   // Fetch usb connected Tessels
-  return usb.findConnections()
-    .each(function(connection){
-      console.log('USB CONNECTION ', connection.serialNumber);
-      return connection;
-    })
-    .then(function(){
-      console.log('Scanning for LAN connections...');
-    })
-    .then(lan.findConnections)
-    .each(function(connection){
-      console.log('LAN CONNECTION ip: ', connection.ip, ', host: ', connection.auth.host, ', Authorized: ', connection.authorized);
-    })
+  console.info('Scanning for connections...');
+  return Promise.join( usb.findConnections(), lan.findConnections(), function(usb, lan){
+        var connections = {
+          usb: usb,
+          lan: lan
+        };
+        for (var i = 0; i < connections.usb.length; i++) {
+          console.log('USB CONNECTION ', connections.usb[i].serialNumber);
+        }
+        for (var i = connections.lan.length - 1; i >= 0; i--) {
+          console.log('LAN CONNECTION ip: ', connections.lan[i].ip, ', host: ', connections.lan[i].auth.host, ', Authorized: ', connections.lan[i].authorized);
+        }
+        return connections;
+    });
 }
 
 /*
