+++ lib/usb_connection.js
@@ -93,11 +93,9 @@ function findConnections() {
       resolve(connections);
     })
     .map(function(conn){
-      console.log(typeof conn);
       return conn.open();
     })
     .catch(function(error, conn){
-      console.log(error, conn.connectionType)
       conn.initError = error;
       return conn;
     })
